src/requirements.txt
djangorestframework==3.13.1
djangorestframework-camel-case==1.1.2
entrypoints==0.3
requests==2.27
gunicorn==20.1.0
django-cors-headers==3.11.0
celery==5.1.2
mysqlclient==2.1.0
django-redis==5.2.0
